Refactor validate_supplier_number into ValidateLitigatorSupplierNumber
