+++ app/validators/claim/litigator_common_validations.rb
@@ -1,5 +1,7 @@
 module Claim
   module LitigatorCommonValidations
+    include ValidateLitigatorSupplierNumber
+
     def self.included(base)
       base.class_eval do
         def self.first_step_common_validations
@@ -41,15 +43,7 @@ module Claim
       validate_on_or_before(Date.today, :case_concluded_at, 'check_not_in_future')
     end
 
-    def validate_supplier_number
-      validate_presence(:supplier_number, 'blank')
-
-      return unless @record.supplier_number.present?
-
-      validate_pattern(:supplier_number, supplier_number_regex, 'invalid')
-      return if @record.errors.key?(:supplier_number)
-      validate_inclusion(:supplier_number, provider_supplier_numbers, 'unknown')
-    end
+    # validate_supplier_number called from ValidateLitigatorSupplierNumber
 
     # local helpers
     # ---------------------------
